// src/ReportGenerator/model/data/TeacherPresentation.ts
export class TeacherPresentation {
  teacher_name_c: string;
  journals_title: string;
  other_authors: string;
  order_authors: string;
  address_authors: string;
  publication: string; // boolean
  seminar: string;
  year_id_s: number|string;
  month_id_s: number|string;
  day_id_s: number|string;
  year_id_e: number|string;
  month_id_e: number|string;
  day_id_e: number|string;
  country: string;
  city: string;
  page_s: string|number;
  page_e: string|number;
  project: string;
}